+++ app/Providers/EventServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace App\Providers;
 
-use Illuminate\Support\Facades\Event;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 
 class EventServiceProvider extends ServiceProvider
@@ -15,7 +14,10 @@ class EventServiceProvider extends ServiceProvider
     protected $listen = [
         \Illuminate\Auth\Events\Registered::class => [
             \Illuminate\Auth\Listeners\SendEmailVerificationNotification::class,
-//            \App\Listeners\SendEmailApprovalNotification::class,
+        ],
+
+        \Illuminate\Auth\Events\Verified::class => [
+            \App\Listeners\SendEmailApprovalNotification::class,
         ],
     ];