+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Log of unreleased changes.
 ### Added
 
 - Requires email verification.
+- Sends `Approval` email when Ho.Re.Ca. user is verified.
 
 ### Fixed